ELOP-452; curvedtrack->point, ID 변경, signal 표시수정

+++ RailmlEditor/Models/RailModel.cs
@@ -72,7 +72,7 @@ namespace RailmlEditor.Models
         public AdditionalName AdditionalName { get; set; }
 
         [XmlAttribute(AttributeName = "pos")]
-        public double Pos { get; set; }
+        public int Pos { get; set; }
 
         // Visual Coordinates
         [XmlAttribute]
